Display Social Circles on Accept/Reject
This change makes use of the callback functions to update the
information being displayed on the view.

Things to remember:
1. Callbacks, once set up are called on the event they are set up for.
For ex: "child_added" events trigger the callback it was set up for.

2. There was no callback for "child_removed", so CirclesPending global
still had the old values which were being displayed by the view. By
adding the "child_removed" callback and deleting the corresponding entry
in the CirclesPending array ensures that the view gets updated with the
new information.

3. Same callback support added for AccountCtrl.

4. Data Binding, you are one asshole

NOTE:- When creating a new group, the information is not being displayed
on the Wallet Tab. This is again because Circles.get() is not called
anywhere to reset the global. Need to figure out a way to do that.