The Big UI change with Google Analytics changes!
This change is a big one, here are a list of things which are going to
change:

1. UI improvements for the settings tab and when the user is within a
circle.
2. When the user clicks a circle, only the users who have accepted the
invite are displayed(Refer to
Circles/CircleMembers/$circleID/AcceptedMembers for the list of users,
this can be used for sending the messages selectively).
3. User can click on another user to get their information.
4. Google Analytics for specific UI clicks are now recorded, this can be
useful for future enhancements or just to understand user behavior.