Fix for badges not updating correctly
The root cause of the problem:
In WalletDetailCtrl the path for displaying data in the "Wallet" tab is
/Users/$userid/AcceptedCircles/Info/$pushref. The ChatCtrl assumes that
the user data for each circle is under the path
Users/$userid/AcceptedCircles/Info/$circleid. So the discrepancy in the
assumption causes the badges counter logic to never work. This change
should fix it.